유저 생성 시 요청 처리

+++ compute-service/src/main/kotlin/com/ttalkak/compute/compute/adapter/in/web/request/UpdateStatusRequest.kt
@@ -4,6 +4,6 @@ data class UpdateStatusRequest(
     val maxCompute: Int,
     val availablePortStart: Int,
     val availablePortEnd: Int,
-    val maxMemory: Double = 0.0,
+    val maxMemory: Int = 0,
     val maxCPU: Double = 0.0,
 )
